move components folder
